Abstracted Instruments to its own component
