+++ src/App.svelte
@@ -1,39 +1,5 @@
 <script lang="ts">
-  import { CircleOfFifths, Guitar, Layout, Piano, ScaleConfig } from '@/components';
-  import { getUID } from '@/utils';
-
-  type InstrumentTypes = 'guitar' | 'piano';
-
-  type Instrument = {
-    id: number;
-    type: InstrumentTypes;
-  };
-
-  let instruments: Instrument[] = [
-    {
-      id: getUID(),
-      type: 'guitar'
-    },
-    {
-      id: getUID(),
-      type: 'piano'
-    }
-  ];
-
-  function addInstrument(type: InstrumentTypes) {
-    instruments = [
-      ...instruments,
-      {
-        id: getUID(),
-        type
-      }
-    ];
-  }
-
-  function removeInstrument(index: number) {
-    instruments.splice(index, 1);
-    instruments = instruments;
-  }
+  import { CircleOfFifths, Instruments, Layout, ScaleConfig } from '@/components';
 </script>
 
 <Layout>
@@ -51,24 +17,7 @@
     <h2>
       Instruments
     </h2>
-    <div>
-      {#each instruments as { id, type }, i (id)}
-        {#if type === 'guitar'}
-          <Guitar />
-        {:else if type === 'piano'}
-          <Piano />
-        {/if}
-        <button on:click={() => removeInstrument(i)}>
-          Remove
-        </button>
-      {/each}
-    </div>
-    <button on:click={() => addInstrument('guitar')}>
-      Add Guitar
-    </button>
-    <button on:click={() => addInstrument('piano')}>
-      Add Piano
-    </button>
+    <Instruments />
   </main>
 </Layout>
