+++ src/components/Instruments/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Instruments.svelte';
