+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export { default as CircleOfFifths } from './CircleOfFifths';
 export { default as Dropdown } from './Dropdown';
 export { default as Guitar } from './Guitar';
+export { default as Instruments } from './Instruments';
 export { default as Layout } from './Layout';
 export { default as Piano } from './Piano';
 export { default as ScaleConfig } from './ScaleConfig';
